+++ FractalsWpf/JuliaSet.cs
@@ -7,7 +7,7 @@ namespace FractalsWpf
 {
     public class JuliaSet : IFractals
     {
-        public int[] CreatePixelArray(
+        public ushort[] CreatePixelArray(
             Complex c,
             Complex c1,
             Complex c2,
@@ -31,15 +31,15 @@ namespace FractalsWpf
             return pixels.ToArray();
         }
 
-        private static int BeginsToDivergeAt(Complex c, Complex z, int maxIterations)
+        private static ushort BeginsToDivergeAt(Complex c, Complex z, int maxIterations)
         {
             foreach (var iterations in Enumerable.Range(0, maxIterations))
             {
                 z = z * z + c;
-                if (z.Real * z.Real + z.Imaginary * z.Imaginary >= 4d) return iterations;
+                if (z.Real * z.Real + z.Imaginary * z.Imaginary >= 4d) return (ushort)iterations;
             }
 
-            return maxIterations;
+            return (ushort)maxIterations;
         }
     }
 }
